requirements.txt
async-lru==2.0.4
async-property==0.2.2
async-timeout==4.0.3
asyncstdlib==3.12.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mplfinance==0.12.10b0
numpy==1.26.4
oauthlib==3.2.2
pandas==2.1.1
pandas-ta==0.3.14b0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-telegram-bot==21.9
requests==2.31.0
requests-mock==1.12.1
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.3
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
